Restyle `check` & `update` buttons (#44)
* Restyle `check` & `update` buttons

It was hard to tell if the `check` or `update` buttons were actually clicked; causing me to click several times just to make sure. This fixes that by using the default Home Assistant animation for a button click. While here, I fixed the got rid of the button floats. This allows the buttons to be displayed in the same order as they are defined in the source.

* revert `paper-button` to `mwc-button`